requirements.txt
streamlit
torch
torchvision
diffusers
transformers
pillow
gtts
accelerate
safetensors
